Change AppError Display to use u16 instead of i32
Updated the Display implementation for AppError to format the error code as a u16 instead of i32, ensuring consistency with the intended error code type.

+++ src-tauri/src/core/errors.rs
@@ -21,7 +21,7 @@ impl Serialize for AppError {
 
 impl std::fmt::Display for AppError {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", *self as i32)
+        write!(f, "{}", *self as u16)
     }
 }
